Clean up Grant and Token models
- Move things into WithExpiration and WithRevocation concerns
