+++ app/models/token.rb
@@ -1,21 +1,11 @@
 class Token < ApplicationRecord
-  belongs_to :grant, required: true
+  include WithExpiration
+  include WithRevocation
 
-  validates :grant, presence: true
-  validates :expires_at, presence: true
+  belongs_to :client, required: true
 
-  def expires_in
-    expires_at - Time.now
-  end
-
-  # @return [Boolean] whether this grant has expired yet
-  def expired?
-    expires_at&.past?
-  end
-
-  def revoked?
-    revoked_at&.past? || grant.revoked?
-  end
+  validates :client, presence: true
+  validates :user_id, presence: true
 
   def acceptable?
     !(expired? || revoked?)
